refac! move dynamo related code to a wrapper
