+++ api/environments/dynamodb/__init__.py
@@ -1,0 +1,1 @@
+from .dynamodb_wrapper import DynamoIdentityWrapper  # noqa
